Add a periodic zones-to-cache function.
A setting is added (zones-to-cache) that is a list of domains=url
pairs.  Periodically (based on the TTLs in the zone) the zonefile
will retrieved from the URL and its contents put into the cache. ATM
only http and https is supported.

Typical use:

	zones-to-cache=.=https://www.internic.net/domain/root.zone

Some restrictions apply:

  - records in a zone file should be grouped by name & type
  - we do not handle multiple SOA in the zone file.
  - sigs corresponding to records should follow those records

Records are considered authoritative if they use the exact zone name
or for DS records if the parent of the name is the exact zone name.

No signatures are validated, that will happen on-demand if the
records are used.

ATM I do not fill the aggressive nsec cache. It would require
validation since it assumes it contents are validated, afaik.