dnsdist: Skip some memory allocations in client mode
dnsdist does not use most of the configuration in client mode, mostly
caring about the settings needed to access the console, but
we still need to parse all directives and create objects due to the
way our Lua parsing works.
We don't, however, need to allocate the whole storage needed for
packet caches, for backend states or for the in-memory ring buffers.
This reduces the memory consumption of a console client to roughly 1 MB
from hundreds of them in some cases.