speedtest: Count the system time as well as the user time
Our tests are designed to run for 100 ms of CPU time, but this is currently
measured with `ITIMER_VIRTUAL` which means only when the process is executing
in userspace. Switching to `ITIMER_PROF` would also account for the time spent
when the system is running on behalf of the process, which seems closer to what
we want.