Created a map of faf sockets (per thread; to specific ComboAddresses; no clean up yet)
Still do not like this approach very much. Part of the complexity comes
from having to handle both replying and non-replying targets for NOD queries.

Also did not look at TCP/DoT yet. I suppose connection pooling for these is out
of the question, unless we *know* beforehand if the target is supposed to reply.