m4: Add option for 64-bit time_t on 32-bit systems with glibc-2.34
glibc-2.34 includes the user-facing part of the 64-bit time_t support
for 32-bit systems. We treat this feature as experimental, as it causes
ABI issues, if 64-bit time_t values are passed to libraries that are
build with 32-bit time_t. It is safe for local use of 64-bit time_t.

Signed-off-by: Sven Wegener <sven.wegener@stealer.net>