dnsdist: Check the size of the query when re-sending a DoH query
When the UDP response to a DoH query was truncated, we re-send the
query via TCP, passing it to a TCP worker. We need to edit the ID
to its original value before that, and while there is no reason
that the query is smaller than a dnsheader, we need to check its size,
not the size of the response.

(cherry picked from commit 5d552300959a5133546bb65ddeb8ce918d598c08)