dnsdist: Build with -fvisibility=hidden by default
Not marking our symbols visible by default allows the compiler to do
better optimizations, especially with Link-Time Optimizations that
several distributions are now enabling by default.
The most obvious optimization comes from dead code removal because
the compiler now knows that it cannot be used by an external shared
object, but better inlining might also occur.
Even without LTO, it reduces the final size of stripped binaries and
might improve loading times a bit.

Note that we already mark the symbols that we want to export with the
'default' visibilty, for example for Lua FFI.