dnsdist: Better code coverage stats when enabled
When exiting on user request, dnsdist calls _Exit() to avoid any
issue caused by the order in which the destructors of our long-lived
objects are called (it would be nice to fix this properly, but it
would require a fairly large refactoring).
Since the static destructors are not called in that case, the most
recent coverage data is not always properly dumped to disk.
Explicitly calling __gcov_dump` before exiting ensures that no data
is lost.