Also collect response stats on packet cache hit.
This has to be evaluated for performance impact.
If the impact is relatively high, we could change the size
histogram to be raw counts (like qtype) and do the transformation
to buckets on reading the data. This would avoid the finding of the right
bucket in the histogram code for each packet processed.

If the performance is still too high after that, we should  at least
document that packet count hits are not accounted in the response
histograms.