Implementation of serve-stale from record cache.
If a resolve fails, we try it again with serveStale is true.  If
serveStale is true or a record is already being server stale, the
record cache is willing to produce (and extend the ttd) of stale
records. It wil also keep a count of the extensions, to be able to
limit those and trigger a task te refresh once every while.

If we (potentially) serve stale, we are less aggessive evicting
stale records from the record cache.

Enable by setting server-stale-extensions (default 0). The unit is
30s. So a value of 2880 will keep serving the record for 24 hours, even
if it cannot be refreshed. If the original ttl of a record is less than 30,
the extension unit will be that ttl.