dnsdist: Set TCP_NODELAY on the TCP connection to backends
Setting `TCP_NODELAY` disables Nagle's algorithm, which is unfortunately
not always playing nice with TCP delayed ACKs. This is especially true
when the backend supports out-of-order processing, allowing dnsdist to
send several queries at once without waiting for an answer. In that case
dnsdist often has to wait several milliseconds (20ms is common) before
the remote end decides to send the delayed ACK which allows us to send
the next query.