rec: Save the last nsspeed recorded plus output it in rec_control dump-nsspeeds
Fixes #11736 in the most basic way. There are some plans to unify
the various NS state tables into a single elaborate data structure,
that would be the moment to store a moving average or similar.

Also modify the output a bit: show times in ms instead of us and
without all the decimals.