Rec: Don't reload Lua config if it hasn't changed
This also groups together 1) the list of frame stream servers, 2) the config from which
the list was created and 3) the config's generation into a single struct called
FrameStreamServersInfo. The struct is used to compare the old and new configuration to
decide whether to destroy the old config object or not.

Part of #11795

(cherry picked from commit afaf1b5d87c4a4961eadaf114855a335711c33c8)