builder: add --no-install-recommends to apt-get install
This was triggered by Ubuntu Kinetic pulling in a version of
systemd-resolved that breaks inside Docker. systemd-resolved
is an indirect (via Recommends somewhere) dependency of devscripts,
which we need.

However, if we were relying on Recommends, that was a bug, so I'm
applying the flag to all distributions.

(cherry picked from commit ecb70be6a176a6630858a3cddaae02446fbb9196)

backport of #11891