rec: Update docs for getting started and compiling.
I'm also making the mention of PowerDNS Recursor more consistent:
no ``the`` and either :program:`PowerDNS Recursor` or :program:`Recursor`.
Use short form if the long form was used already in a section or
it would become tedious to have the long form for another reason.

Fixes #11828, though the recursor.conf man page is not going to
happen anytime soon, as we lack the infrastructure for that and
maintaining both a .rst file and a man page would be a pain.

With released versions, pdns_recursor --config=default helps a bit.
In master pdns_recursor --config=[default|check|diff] was added.