If we encounter a loop in QM, continue with the next iteration.
We would not make progres anyway, and end up repeatedly asking the full name
from the roots as the "Trying less specific NS" code path is hit.

Noted by and fix from  @phonedph1

Fixes #12090

As this is an optimzation for broken domains, this is not an immediate candidate
for backporting, though it should be considered for that after more testing.