Convert DomainInfo::DomainKind to ZoneKind class-wrapped-enum
This patch 'lifts' DomainKind out of DomainInfo (so that it can be used
when no DomainInfo object is available), renames it to ZoneKind, and
moves its conversion and classification functions into the new class
as member functions.

This was inspired by #12205, where code that did not have a DomainInfo
object was unable to make use of the isSecondary() classification
function, which resulted in a logic error for zones of type 'Consumer'.