Only store NSEC3 records in aggressive cache if we expect them to be effective.
The aggressive cache is not very effective for large NSEC3 domains: each NSEC3 record only
covers a relatively small amount of random names.

We only want to put NSEC3 records in the aggresive cache if there is a decent
chance a hash will hit it. Hashes are random wrt the corresponding names.

So look at the bitwise common prefix of the owner and next, and do
not store the record if the common prefix is long.

We also might want to introduce a switch to completely forget the aggressive
cache for the NSEC3 case.