Sanitize d_orig_ttl
The computed orig_ttl can be wrong if d_now.tv_sec is more recent
than the d_now.tv_sec the ttd was computed from. This can happen
if we went out to get e.g. keys and that took time. d_orig_ttl wraps
around in that case if the TTL value was smaller than that delay.
Work around that and make sure d_orig_ttl is within the legal range.